+++ example/src/l18n/intl_config.ts
@@ -1,4 +1,5 @@
 import KTextConstants from "@/shared/constants/variables/text_constants";
+import { getCloudflareContext } from "@opennextjs/cloudflare";
 import { setIntlConfig } from "cloudflare-next-intl";
 
 declare global {
@@ -8,5 +9,9 @@ declare global {
 const intlConfig = setIntlConfig({
     locales: KTextConstants.locales,
     defaultLocale: KTextConstants.defaultLocale,
+    cookieConsent: {
+        privacyPolicyDate: "2026-01-01",
+        getCloudflareContext: getCloudflareContext,
+    },
 });
 export default intlConfig;